requirements.txt
discord.py~=1.4.1
dateparser~=1.0.0
parsedatetime~=2.6
pytz~=2021.1
python-dateutil~=2.8.1
tortoise-orm~=0.16.21
pony~=0.7.14
